[arrow-cast] Improve support for casting string view to numeric
